+++ packages/@tailwindcss-upgrade/src/codemods/template/migrate-arbitrary-value-to-bare-value.test.ts
@@ -34,7 +34,6 @@ test.each([
 
   // Leading is special, because `leading-[123]` is the direct value of 123, but
   // `leading-123` maps to `calc(--spacing(123))`.
-  ['leading-[1]', 'leading-none'],
   ['leading-[123]', 'leading-[123]'],
 
   ['data-[selected]:flex', 'data-selected:flex'],
@@ -60,7 +59,7 @@ test.each([
     'data-[selected]:aria-[selected="true"]:aspect-[12/34]',
     'data-selected:aria-selected:aspect-12/34',
   ],
-])('%s => %s', async (candidate, result) => {
+])('%s => %s (%#)', async (candidate, result) => {
   let designSystem = await __unstable__loadDesignSystem('@import "tailwindcss";', {
     base: __dirname,
   })